Added some more project logic.
Projectfactory and projectLi creator. Both of these can be collapsed
into a function with the todo counterpart.

Need to refactor as project and todo factory are almost identical,
raise the abstraction up a level and do both in one.
As with the element creator.

Can add projects from array into the sidebar. Now need to do this
dynamically and add a seperate page for each project.

Dynamically create as I did in restaurant app.