Added wrappers for addtodo,canceltodo and newTodo
This fixed the issue where the event listener would not fire if you
clicked towards the outside of the button. Couldn't find anything On
google, assumed it was due to the button transforming when active.

Added the event listener to the wrapper instead, which doesnt transform.
This seems to have fixed the issue. Win!